pyproject.toml
# pyproject.toml - Project configuration for prevent-outage-edge-testing (POET)
# Compatible with both Poetry and uv package managers.
#
# CLI entrypoint: `poet` (Prevent Outage Edge Testing)

[project]
name = "prevent-outage-edge-testing"
version = "0.1.0"
description = "Knowledge pack library + AI-assisted builder for edge reliability testing"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Edge Reliability Team" }]
keywords = ["testing", "reliability", "edge", "observability", "chaos-engineering", "fault-injection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "pydantic>=2.5.0",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "httpx>=0.26.0",
    "jinja2>=3.1.2",
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0.0",
]
# For systems that support privileged extractors (Linux with eBPF, macOS with DTrace)
privileged = [
    "bcc>=0.29.0;sys_platform=='linux'",
]

[project.scripts]
poet = "prevent_outage_edge_testing.cli.main:app"

[project.urls]
Homepage = "https://github.com/example/prevent-outage-edge-testing"
Documentation = "https://github.com/example/prevent-outage-edge-testing#readme"
Repository = "https://github.com/example/prevent-outage-edge-testing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prevent_outage_edge_testing"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "privileged: tests requiring elevated permissions (DTrace, eBPF)",
    "slow: tests that take significant time",
    "integration: integration tests requiring external services",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "PT"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["src/prevent_outage_edge_testing"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
